Add Windows version comments to the python manifest.
This is so it becomes easier to add or remove future values as time passes and when support for older versions of Windows gets dropped from python. MSVC's vNEXT (V18) will soon drop all but Windows 8.1 so if and when Python migrates to that version of MSVC that means support for Windows older than 8.1 would need to be dropped then.